requirements.txt
beanstalkc3>=0.4.0
couchbase==2.5.12
marshmallow>=3.0.5
psutil
pyyaml
tornado>=4.5
mock
pytest
pytest-cov
pytest-tornado
